added new nook - useReviewChanges
