+++ src/custom-hooks/use-review-changes.tsx
@@ -1,0 +1,40 @@
+import { useState } from 'react';
+import { EMPTY_LINE, EMPTY_RATING } from '../const';
+import { FormEvent } from 'react';
+
+function useReviewChanges () {
+  const [review, setReview] = useState({
+    comment: EMPTY_LINE,
+    rating: EMPTY_RATING,
+  });
+
+  function choseStar(event: React.MouseEvent<HTMLInputElement>) {
+    const stars = parseInt((event.target as HTMLInputElement).value, 10);
+    setReview({
+      ...review,
+      rating: stars,
+    });
+  }
+
+  function changeText(event: React.ChangeEvent<HTMLTextAreaElement>){
+    const text: string = event.target.value;
+    setReview({
+      ...review,
+      comment: text,
+    });
+  }
+
+  function resetData (event: FormEvent<HTMLFormElement>) {
+    setReview({
+      ...review,
+      comment: EMPTY_LINE,
+      rating: EMPTY_RATING,
+    });
+    event.currentTarget.reset();
+  }
+
+  return [changeText, resetData, choseStar, review];
+}
+
+
+export default useReviewChanges;
